The app can change the speed rate when the configuration is updated.

+++ app/src/main/java/ar/com/westsoft/listening/data/datasource/DictGameSettingsDSO.kt
@@ -3,5 +3,5 @@ package ar.com.westsoft.listening.data.datasource
 data class DictGameSettingsDSO(
     val readWordAfterCursor: Int,
     val readWordBeforeCursor: Int,
-    val speechRate: Double
+    val speechRate: Float
 )
